web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = "The Packet Wizards Blog"
authors = ["Leonardo Monteiro <decastromonteiro@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
asgiref="^3.2.7"
Django="^3.0.7"
django-markdownx="^3.0.1"
Markdown="^3.2.2"
Pillow="^7.1.2"
Pygments="^2.6.1"
pytz="^2020.1"
sqlparse="^0.3.1"
gunicorn="^20.0.4"
psycopg2-binary="^2.8.5"
celery = "^5.0.0"
tweepy = "^3.9.0"
amqp = "^5.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
